+++ packages/auth/src/tenancy/tenancy.js
@@ -73,92 +73,6 @@ exports.tryAddTenant = async (tenantId, userId, email) => {
   await Promise.all(promises)
 }
 
-const DocumentTypes = {
-  USER: "us",
-}
-const UNICODE_MAX = "\ufff0"
-
-/**
- * Gets parameters for retrieving users.
- * Duplicate of "../db/utils" due to circular dependency
- */
-const getGlobalUserParams = (globalId, otherProps = {}) => {
-  if (!globalId) {
-    globalId = ""
-  }
-  return {
-    ...otherProps,
-    startkey: `${DocumentTypes.USER}${SEPARATOR}${globalId}`,
-    endkey: `${DocumentTypes.USER}${SEPARATOR}${globalId}${UNICODE_MAX}`,
-  }
-}
-
-const removeTenantFromInfoDB = async tenantId => {
-  try {
-    const infoDb = getDB(PLATFORM_INFO_DB)
-    let tenants = await infoDb.get(TENANT_DOC)
-    tenants.tenantIds = tenants.tenantIds.filter(id => id !== tenantId)
-
-    await infoDb.put(tenants)
-  } catch (err) {
-    console.error(`Error removing tenant ${tenantId} from info db`, err)
-    throw err
-  }
-}
-
-const removeUsersFromInfoDB = async tenantId => {
-  try {
-    const globalDb = exports.getGlobalDB(tenantId)
-    const infoDb = getDB(PLATFORM_INFO_DB)
-    const allUsers = await globalDb.allDocs(
-      getGlobalUserParams(null, {
-        include_docs: true,
-      })
-    )
-    const allEmails = allUsers.rows.map(row => row.doc.email)
-    // get the id docs
-    let keys = allUsers.rows.map(row => row.id)
-    // and the email docs
-    keys = keys.concat(allEmails)
-    // retrieve the docs and delete them
-    const userDocs = await infoDb.allDocs({
-      keys,
-      include_docs: true,
-    })
-    const toDelete = userDocs.rows.map(row => {
-      return {
-        ...row.doc,
-        _deleted: true,
-      }
-    })
-    await infoDb.bulkDocs(toDelete)
-  } catch (err) {
-    console.error(`Error removing tenant ${tenantId} users from info db`, err)
-    throw err
-  }
-}
-
-const removeGlobalDB = async tenantId => {
-  try {
-    const globalDb = exports.getGlobalDB(tenantId)
-    await globalDb.destroy()
-  } catch (err) {
-    console.error(`Error removing tenant ${tenantId} users from info db`, err)
-    throw err
-  }
-}
-
-const removeTenantApps = async () => {
-  // TODO
-}
-
-exports.deleteTenant = async tenantId => {
-  await removeTenantFromInfoDB(tenantId)
-  await removeUsersFromInfoDB(tenantId)
-  await removeGlobalDB(tenantId)
-  await removeTenantApps(tenantId)
-}
-
 exports.getGlobalDB = (tenantId = null) => {
   // tenant ID can be set externally, for example user API where
   // new tenants are being created, this may be the case
